+++ webapi/src/test/scala/org/knora/webapi/e2e/v2/ResponseCheckerV2Spec.scala
@@ -18,7 +18,7 @@ class ResponseCheckerV2Spec extends CoreSpec() {
   "ResponseCheckerV2" should {
     "not throw an exception if received and expected resource responses are the same" in {
       val expectedAnswerJSONLD =
-        FileUtil.readTextFile(Paths.get("test_data/resourcesR2RV2/ThingWithLinkComplex.jsonld"))
+        FileUtil.readTextFile(Paths.get("..", "test_data/resourcesR2RV2/ThingWithLinkComplex.jsonld"))
 
       ResponseCheckerV2.compareJSONLDForResourcesResponse(
         expectedJSONLD = expectedAnswerJSONLD,
@@ -28,7 +28,7 @@ class ResponseCheckerV2Spec extends CoreSpec() {
 
     "not throw an exception if received and expected mapping responses are the same" in {
       val expectedAnswerJSONLD =
-        FileUtil.readTextFile(Paths.get("test_data/standoffR2RV2/mappingCreationResponse.jsonld"))
+        FileUtil.readTextFile(Paths.get("..", "test_data/standoffR2RV2/mappingCreationResponse.jsonld"))
 
       ResponseCheckerV2.compareJSONLDForMappingCreationResponse(
         expectedJSONLD = expectedAnswerJSONLD,
@@ -38,8 +38,8 @@ class ResponseCheckerV2Spec extends CoreSpec() {
 
     "throw an exception if received and expected resource responses are different" in {
       val expectedAnswerJSONLD =
-        FileUtil.readTextFile(Paths.get("test_data/resourcesR2RV2/ThingWithLinkComplex.jsonld"))
-      val receivedAnswerJSONLD = FileUtil.readTextFile(Paths.get("test_data/resourcesR2RV2/ThingWithListValue.jsonld"))
+        FileUtil.readTextFile(Paths.get("..", "test_data/resourcesR2RV2/ThingWithLinkComplex.jsonld"))
+      val receivedAnswerJSONLD = FileUtil.readTextFile(Paths.get("..", "test_data/resourcesR2RV2/ThingWithListValue.jsonld"))
 
       assertThrows[AssertionError] {
         ResponseCheckerV2.compareJSONLDForResourcesResponse(
@@ -51,9 +51,9 @@ class ResponseCheckerV2Spec extends CoreSpec() {
 
     "throw an exception if the values of the received and expected resource responses are different" in {
       val expectedAnswerJSONLD =
-        FileUtil.readTextFile(Paths.get("test_data/resourcesR2RV2/BookReiseInsHeiligeLand.jsonld"))
+        FileUtil.readTextFile(Paths.get("..", "test_data/resourcesR2RV2/BookReiseInsHeiligeLand.jsonld"))
       val receivedAnswerJSONLD =
-        FileUtil.readTextFile(Paths.get("test_data/resourcesR2RV2/BookReiseInsHeiligeLandPreview.jsonld"))
+        FileUtil.readTextFile(Paths.get("..", "test_data/resourcesR2RV2/BookReiseInsHeiligeLandPreview.jsonld"))
 
       assertThrows[AssertionError] {
         ResponseCheckerV2.compareJSONLDForResourcesResponse(
@@ -65,10 +65,10 @@ class ResponseCheckerV2Spec extends CoreSpec() {
 
     "throw an exception if the number of values of the received and expected resource responses are different" in {
       val expectedAnswerJSONLD =
-        FileUtil.readTextFile(Paths.get("test_data/resourcesR2RV2/NarrenschiffFirstPage.jsonld"))
+        FileUtil.readTextFile(Paths.get("..", "test_data/resourcesR2RV2/NarrenschiffFirstPage.jsonld"))
       // number of StillImageFileValue is wrong
       val receivedAnswerJSONLD =
-        FileUtil.readTextFile(Paths.get("test_data/responseCheckerR2RV2/NarrenschiffFirstPageWrong.jsonld"))
+        FileUtil.readTextFile(Paths.get("..", "test_data/responseCheckerR2RV2/NarrenschiffFirstPageWrong.jsonld"))
 
       assertThrows[AssertionError] {
         ResponseCheckerV2.compareJSONLDForResourcesResponse(
@@ -80,9 +80,9 @@ class ResponseCheckerV2Spec extends CoreSpec() {
 
     "throw an exception if received and expected mapping responses are different" in {
       val expectedAnswerJSONLD =
-        FileUtil.readTextFile(Paths.get("test_data/standoffR2RV2/mappingCreationResponse.jsonld"))
+        FileUtil.readTextFile(Paths.get("..", "test_data/standoffR2RV2/mappingCreationResponse.jsonld"))
       val receivedAnswerJSONLD =
-        FileUtil.readTextFile(Paths.get("test_data/standoffR2RV2/mappingCreationResponseWithDifferentLabel.jsonld"))
+        FileUtil.readTextFile(Paths.get("..", "test_data/standoffR2RV2/mappingCreationResponseWithDifferentLabel.jsonld"))
 
       assertThrows[AssertionError] {
         ResponseCheckerV2.compareJSONLDForMappingCreationResponse(