+++ webapi/src/test/scala/org/knora/webapi/messages/util/rdf/RdfFormatUtilSpec.scala
@@ -85,7 +85,7 @@ abstract class RdfFormatUtilSpec(featureToggle: FeatureToggle) extends CoreSpec 
 
   "RdfFormatUtil" should {
     "parse RDF in Turtle format, producing an RdfModel, then format it as Turtle again" in {
-      val inputTurtle: String = FileUtil.readTextFile(Paths.get("test_data/resourcesR2RV2/BookReiseInsHeiligeLand.ttl"))
+      val inputTurtle: String = FileUtil.readTextFile(Paths.get("..", "test_data/resourcesR2RV2/BookReiseInsHeiligeLand.ttl"))
       val inputModel: RdfModel = rdfFormatUtil.parseToRdfModel(rdfStr = inputTurtle, rdfFormat = Turtle)
       checkModelForRdfTypeBook(inputModel)
 
@@ -97,7 +97,7 @@ abstract class RdfFormatUtilSpec(featureToggle: FeatureToggle) extends CoreSpec 
 
     "parse RDF in JSON-LD format, producing an RdfModel, then format it as JSON-LD again" in {
       val inputTurtle: String =
-        FileUtil.readTextFile(Paths.get("test_data/resourcesR2RV2/BookReiseInsHeiligeLand.jsonld"))
+        FileUtil.readTextFile(Paths.get("..", "test_data/resourcesR2RV2/BookReiseInsHeiligeLand.jsonld"))
       val inputModel: RdfModel = rdfFormatUtil.parseToRdfModel(rdfStr = inputTurtle, rdfFormat = JsonLD)
       checkModelForRdfTypeBook(inputModel)
 
@@ -108,7 +108,7 @@ abstract class RdfFormatUtilSpec(featureToggle: FeatureToggle) extends CoreSpec 
     }
 
     "parse RDF in Turtle format, producing a JsonLDDocument, then format it as Turtle again" in {
-      val inputTurtle: String = FileUtil.readTextFile(Paths.get("test_data/resourcesR2RV2/BookReiseInsHeiligeLand.ttl"))
+      val inputTurtle: String = FileUtil.readTextFile(Paths.get("..", "test_data/resourcesR2RV2/BookReiseInsHeiligeLand.ttl"))
       val inputModel: RdfModel = rdfFormatUtil.parseToRdfModel(rdfStr = inputTurtle, rdfFormat = Turtle)
       val inputJsonLDDocument: JsonLDDocument =
         rdfFormatUtil.parseToJsonLDDocument(rdfStr = inputTurtle, rdfFormat = Turtle)
@@ -125,7 +125,7 @@ abstract class RdfFormatUtilSpec(featureToggle: FeatureToggle) extends CoreSpec 
     }
 
     "parse RDF in RDF/XML format, producing a JsonLDDocument, then format it as RDF/XML again" in {
-      val inputRdfXml: String = FileUtil.readTextFile(Paths.get("test_data/resourcesR2RV2/BookReiseInsHeiligeLand.rdf"))
+      val inputRdfXml: String = FileUtil.readTextFile(Paths.get("..", "test_data/resourcesR2RV2/BookReiseInsHeiligeLand.rdf"))
       val inputModel: RdfModel = rdfFormatUtil.parseToRdfModel(rdfStr = inputRdfXml, rdfFormat = RdfXml)
       val inputJsonLDDocument: JsonLDDocument =
         rdfFormatUtil.parseToJsonLDDocument(rdfStr = inputRdfXml, rdfFormat = RdfXml)
@@ -143,14 +143,14 @@ abstract class RdfFormatUtilSpec(featureToggle: FeatureToggle) extends CoreSpec 
 
     "parse RDF in TriG format" in {
       val graphIri = "http://example.org/data#"
-      val inputTrig = FileUtil.readTextFile(Paths.get("test_data/rdfFormatUtil/BookReiseInsHeiligeLand.trig"))
+      val inputTrig = FileUtil.readTextFile(Paths.get("..", "test_data/rdfFormatUtil/BookReiseInsHeiligeLand.trig"))
       val inputModel: RdfModel = rdfFormatUtil.parseToRdfModel(rdfStr = inputTrig, rdfFormat = TriG)
       checkModelForRdfTypeBook(rdfModel = inputModel, context = Some(graphIri))
     }
 
     "parse RDF in N-Quads format" in {
       val graphIri = "http://example.org/data#"
-      val inputTrig = FileUtil.readTextFile(Paths.get("test_data/rdfFormatUtil/BookReiseInsHeiligeLand.nq"))
+      val inputTrig = FileUtil.readTextFile(Paths.get("..", "test_data/rdfFormatUtil/BookReiseInsHeiligeLand.nq"))
       val inputModel: RdfModel = rdfFormatUtil.parseToRdfModel(rdfStr = inputTrig, rdfFormat = NQuads)
       checkModelForRdfTypeBook(rdfModel = inputModel, context = Some(graphIri))
     }
@@ -158,7 +158,7 @@ abstract class RdfFormatUtilSpec(featureToggle: FeatureToggle) extends CoreSpec 
     "read Turtle, add a graph IRI to it, write it to a TriG file, and read back the TriG file" in {
       val graphIri = "http://example.org/data#"
       val rdfSource = RdfInputStreamSource(
-        new BufferedInputStream(Files.newInputStream(Paths.get("test_data/resourcesR2RV2/BookReiseInsHeiligeLand.ttl")))
+        new BufferedInputStream(Files.newInputStream(Paths.get("..", "test_data/resourcesR2RV2/BookReiseInsHeiligeLand.ttl")))
       )
       val outputFile: Path = Files.createTempFile("test", ".trig")
 
@@ -176,7 +176,7 @@ abstract class RdfFormatUtilSpec(featureToggle: FeatureToggle) extends CoreSpec 
     "read Turtle, add a graph IRI to it, write it to an N-Quads file, and read back the N-Quads file" in {
       val graphIri = "http://example.org/data#"
       val rdfSource = RdfInputStreamSource(
-        new BufferedInputStream(Files.newInputStream(Paths.get("test_data/resourcesR2RV2/BookReiseInsHeiligeLand.ttl")))
+        new BufferedInputStream(Files.newInputStream(Paths.get("..", "test_data/resourcesR2RV2/BookReiseInsHeiligeLand.ttl")))
       )
       val outputFile: Path = Files.createTempFile("test", ".trig")
 
@@ -193,7 +193,7 @@ abstract class RdfFormatUtilSpec(featureToggle: FeatureToggle) extends CoreSpec 
 
     "parse RDF in JSON-LD format, producing a JsonLDDocument, then format it as JSON-LD again" in {
       val inputTurtle: String =
-        FileUtil.readTextFile(Paths.get("test_data/resourcesR2RV2/BookReiseInsHeiligeLand.jsonld"))
+        FileUtil.readTextFile(Paths.get("..", "test_data/resourcesR2RV2/BookReiseInsHeiligeLand.jsonld"))
       val inputJsonLDDocument: JsonLDDocument =
         rdfFormatUtil.parseToJsonLDDocument(rdfStr = inputTurtle, rdfFormat = JsonLD)
       checkJsonLDDocumentForRdfTypeBook(inputJsonLDDocument)
@@ -207,7 +207,7 @@ abstract class RdfFormatUtilSpec(featureToggle: FeatureToggle) extends CoreSpec 
 
     "use prefixes and custom datatypes" in {
       val inputJsonLD: String =
-        FileUtil.readTextFile(Paths.get("test_data/resourcesR2RV2/BookReiseInsHeiligeLandSimple.jsonld"))
+        FileUtil.readTextFile(Paths.get("..", "test_data/resourcesR2RV2/BookReiseInsHeiligeLandSimple.jsonld"))
       val inputJsonLDDocument: JsonLDDocument = JsonLDUtil.parseJsonLD(inputJsonLD)
       val outputModel: RdfModel = inputJsonLDDocument.toRdfModel(rdfModelFactory)
 
@@ -231,7 +231,7 @@ abstract class RdfFormatUtilSpec(featureToggle: FeatureToggle) extends CoreSpec 
 
     "parse RDF from a stream and process it using an RdfStreamProcessor" in {
       val inputStream =
-        new BufferedInputStream(Files.newInputStream(Paths.get("test_data/ontologies/anything-onto.ttl")))
+        new BufferedInputStream(Files.newInputStream(Paths.get("..", "test_data/ontologies/anything-onto.ttl")))
       val testStreamProcessor = new TestStreamProcessor
 
       rdfFormatUtil.parseWithStreamProcessor(
@@ -248,7 +248,7 @@ abstract class RdfFormatUtilSpec(featureToggle: FeatureToggle) extends CoreSpec 
       // to a ByteArrayOutputStream.
 
       val fileInputStream =
-        new BufferedInputStream(Files.newInputStream(Paths.get("test_data/ontologies/anything-onto.ttl")))
+        new BufferedInputStream(Files.newInputStream(Paths.get("..", "test_data/ontologies/anything-onto.ttl")))
       val byteArrayOutputStream = new ByteArrayOutputStream()
 
       val formattingStreamProcessor = rdfFormatUtil.makeFormattingStreamProcessor(
@@ -280,7 +280,7 @@ abstract class RdfFormatUtilSpec(featureToggle: FeatureToggle) extends CoreSpec 
 
     "stream RDF data from an InputStream into an RdfModel, then into an OutputStream, then back into an RdfModel" in {
       val fileInputStream =
-        new BufferedInputStream(Files.newInputStream(Paths.get("test_data/ontologies/anything-onto.ttl")))
+        new BufferedInputStream(Files.newInputStream(Paths.get("..", "test_data/ontologies/anything-onto.ttl")))
       val rdfModel: RdfModel = rdfFormatUtil.inputStreamToRdfModel(inputStream = fileInputStream, rdfFormat = Turtle)
       fileInputStream.close()
       assert(rdfModel.contains(expectedThingLabelStatement))